refactor(storage): split blocked capture coordination
