+++ packages/remnic-core/src/storage/tombstone-blocked-capture-index.ts
@@ -1,9 +1,8 @@
 import { execFile } from "node:child_process";
-import { AsyncLocalStorage } from "node:async_hooks";
-import { createHash, randomUUID } from "node:crypto";
-import { mkdir, open as openFile, readFile, readdir, rename, stat, unlink, writeFile } from "node:fs/promises";
+import { randomUUID } from "node:crypto";
+import { mkdir, readFile, readdir, rename, stat, unlink, writeFile } from "node:fs/promises";
 import path from "node:path";
-import { isDeepStrictEqual, promisify } from "node:util";
+import { promisify } from "node:util";
 import { log } from "../logger.js";
 import type { MemoryFile, MemoryFrontmatter } from "../types.js";
 import { RECALL_FALLBACK_DIRS } from "../utils/category-dir.js";
@@ -15,44 +14,30 @@ import {
   type ContentHashIndexLockOptions,
 } from "./content-hash-index.js";
 import {
-  isQueuedReviewFrontmatter, isQueuedReviewMemory, tombstoneBlocked,
-  runTombstoneBlockedChunkMutation,
-  runTombstoneBlockedMutation,
-  type TombstoneBlockedMutationHost,
+  buildExplicitCaptureDedupKey,
+  TombstoneBlockedCaptureWriteLock,
 } from "./tombstone-blocked-capture-mutation.js";
-import { writeMaybeEncryptedFileFromChunks } from "../secure-store/secure-fs.js";
+import type { OfflineSyncMemoryParser } from "./tombstone-blocked-capture-sync.js";
+export {
+  buildExplicitCaptureDedupKey,
+  isTombstoneBlockedCaptureWriteLockBusy,
+  normalizeExplicitCaptureContent,
+} from "./tombstone-blocked-capture-mutation.js";
+export {
+  parseTombstoneBlockedOfflineSyncMemory,
+  type OfflineSyncMemoryParser,
+  TombstoneBlockedCaptureIndexHost,
+} from "./tombstone-blocked-capture-sync.js";
 
 const REBUILD_MAX_ATTEMPTS = 3;
 const execFileAsync = promisify(execFile);
 const REBUILD_RETRY_BASE_MS = 50;
 
-const TOMBSTONE_CAPTURE_WRITE_LOCK_STALE_MS = 60_000;
-const CAPTURE_WRITE_LOCK_MAX_ATTEMPTS = 3;
-const CAPTURE_WRITE_LOCK_RETRY_BASE_MS = 25;
 const ABANDONED_MARKER_MIN_AGE_MS = 60_000;
 const MARKER_PROCESS_START_TOLERANCE_MS = 2_000;
 const DOTNET_UNIX_EPOCH_TICKS = 621_355_968_000_000_000;
 const MARKER_PROCESS_STARTED_AT_MS = Date.now() - process.uptime() * 1000;
-const CAPTURE_WRITE_LOCK_BUSY_MESSAGE = "tombstone-blocked capture write lock remained busy";
 
-const OFFLINE_SYNC_BUFFER_LIMIT_BYTES = 1_048_576;
-const OFFLINE_SYNC_STAGE_READ_BYTES = 64 * 1024;
-const OFFLINE_SYNC_FRONTMATTER_LIMIT_BYTES = 128 * 1024;
-
-type StreamedOfflineSyncIdentity = {
-  identityHash: string;
-};
-
-type PreparedOfflineSyncChunks = {
-  after: MemoryFile | null;
-  chunks: AsyncIterable<Buffer>;
-  stagePath?: string;
-  streamedIdentity?: StreamedOfflineSyncIdentity;
-};
-
-export function isTombstoneBlockedCaptureWriteLockBusy(error: unknown): boolean {
-  return error instanceof Error && error.message === CAPTURE_WRITE_LOCK_BUSY_MESSAGE;
-}
 
 type RebuildMarker = {
   path: string;
@@ -64,115 +49,6 @@ type RebuildMarker = {
   malformed?: boolean;
 };
 
-/**
- * Normalize the content identity used by explicit-capture duplicate checks.
- * This deliberately matches the capture path's historical comparison rather
- * than the punctuation-stripping fact-hash index: the index is only an
- * optimization, while this key gates whether a durable row needs inspection.
- */
-export function normalizeExplicitCaptureContent(value: string): string {
-  return value.toLowerCase().replace(/\s+/g, " ").trim();
-}
-
-/**
- * Build the targeted identity for a tombstone-blocked explicit-capture row.
- * Content uses the same whitespace-only normalization as durable duplicate
- * checks; category and connector preserve their exact trimmed values.
- */
-export function buildExplicitCaptureDedupKey(
-  content: string,
-  category: string,
-  sourceConnector: string | undefined
-): string {
-  const encode = (label: string, value: string): string => `${label} ${value.length} ${value}`;
-  return [
-    encode("category", category),
-    encode("connector", sourceConnector?.trim() ?? ""),
-    encode("content", normalizeExplicitCaptureContent(content)),
-  ].join(" ");
-}
-
-function captureIdentityHash(memory: MemoryFile): string {
-  return createHash("sha256")
-    .update(buildExplicitCaptureDedupKey(memory.content, memory.frontmatter.category, memory.frontmatter.sourceConnector))
-    .digest("hex");
-}
-
-function createExplicitContentNormalizer(onChunk: (chunk: string) => void): {
-  push: (text: string) => void;
-  finish: () => void;
-} {
-  let started = false;
-  let pendingSpace = false;
-  let output = "";
-  const flush = () => {
-    if (output.length > 0) {
-      onChunk(output);
-      output = "";
-    }
-  };
-  return {
-    push(text) {
-      for (const character of text.toLowerCase()) {
-        if (/\s/u.test(character)) {
-          if (started) pendingSpace = true;
-          continue;
-        }
-        if (pendingSpace) {
-          output += " ";
-          pendingSpace = false;
-        }
-        output += character;
-        started = true;
-        if (output.length >= OFFLINE_SYNC_STAGE_READ_BYTES) flush();
-      }
-    },
-    finish: flush,
-  };
-}
-
-async function forEachStagedTextChunk(
-  stagePath: string,
-  offset: number,
-  onText: (text: string) => void
-): Promise<void> {
-  const handle = await openFile(stagePath, "r");
-  const buffer = Buffer.alloc(OFFLINE_SYNC_STAGE_READ_BYTES);
-  const decoder = new TextDecoder("utf-8");
-  let position = offset;
-  try {
-    for (;;) {
-      const { bytesRead } = await handle.read(buffer, 0, buffer.length, position);
-      if (bytesRead === 0) break;
-      position += bytesRead;
-      const text = decoder.decode(buffer.subarray(0, bytesRead), { stream: true });
-      if (text.length > 0) onText(text);
-    }
-    const finalText = decoder.decode();
-    if (finalText.length > 0) onText(finalText);
-  } finally {
-    await handle.close();
-  }
-}
-
-export type OfflineSyncMemoryParser = (filePath: string, content: Buffer) => MemoryFile | null;
-
-export function parseTombstoneBlockedOfflineSyncMemory(
-  filePath: string,
-  content: Buffer,
-  memoryDir: string,
-  parseFrontmatter: (raw: string) => { frontmatter: MemoryFrontmatter; content: string } | null,
-  normalizeFrontmatterForPath: (frontmatter: MemoryFrontmatter, pathRel: string, content?: string) => MemoryFrontmatter,
-  toMemoryPathRel: (baseDir: string, filePath: string) => string
-): MemoryFile | null {
-  const parsed = parseFrontmatter(content.toString("utf8"));
-  if (!parsed) return null;
-  return {
-    path: filePath,
-    frontmatter: normalizeFrontmatterForPath(parsed.frontmatter, toMemoryPathRel(memoryDir, filePath), parsed.content),
-    content: parsed.content,
-  };
-}
 
 export type TombstoneBlockedCaptureIndexOptions = {
   readonly stateDir: string;
@@ -199,9 +75,11 @@ export class TombstoneBlockedCaptureIndex {
   private loadPromise: Promise<ContentHashIndex> | null = null;
   private refreshPromise: Promise<ContentHashIndex> | null = null;
   private authoritative = false;
+  private readonly captureWriteLock: TombstoneBlockedCaptureWriteLock;
 
   constructor(options: TombstoneBlockedCaptureIndexOptions) {
     this.options = options;
+    this.captureWriteLock = new TombstoneBlockedCaptureWriteLock(options);
   }
 
   private createIndex(): ContentHashIndex {
@@ -485,66 +363,12 @@ export class TombstoneBlockedCaptureIndex {
     return this.loadPromise;
   }
 
-  private captureWriteLockPathForHash(identityHash: string): string {
-    const filename =
-      identityHash.length === 0
-        ? "explicit-capture-write.lock"
-        : `explicit-capture-write-${identityHash}.lock`;
-    return path.join(this.options.stateDir, "tombstone-blocked-capture", filename);
-  }
-
-  private captureWriteLockPath(identity: string): string {
-    return this.captureWriteLockPathForHash(
-      identity.length === 0 ? "" : createHash("sha256").update(identity).digest("hex")
-    );
-  }
-
-  private async withSingleCaptureWriteLock<T>(lockPath: string, task: () => Promise<T>): Promise<T> {
-    await mkdir(path.dirname(lockPath), { recursive: true });
-    const retry = Symbol("retry");
-    const permanentFailure = Symbol("permanent-failure");
-    const runWithHeldFileLock = this.options.withHeldFileLock ?? withHeldFileLock;
-    for (let attempt = 0; attempt < CAPTURE_WRITE_LOCK_MAX_ATTEMPTS; attempt += 1) {
-      const result = await runWithHeldFileLock<T | typeof retry | typeof permanentFailure>(
-        lockPath,
-        {
-          staleMs: TOMBSTONE_CAPTURE_WRITE_LOCK_STALE_MS,
-          maxWaitMs: 1_000,
-          pollMs: 25,
-        },
-        async (acquired, controller) => {
-          if (acquired) return await task();
-          return controller.failure === "error" ? permanentFailure : retry;
-        }
-      );
-      if (result === permanentFailure) {
-        throw new Error("tombstone-blocked capture write lock acquisition failed");
-      }
-      if (result !== retry) return result;
-      if (attempt === CAPTURE_WRITE_LOCK_MAX_ATTEMPTS - 1) break;
-      const delayState = Promise.withResolvers<void>();
-      setTimeout(delayState.resolve, CAPTURE_WRITE_LOCK_RETRY_BASE_MS * 2 ** attempt);
-      await delayState.promise;
-    }
-    throw new Error(CAPTURE_WRITE_LOCK_BUSY_MESSAGE);
-  }
-
   async withCaptureWriteLock<T>(task: () => Promise<T>, identity?: string | readonly string[]): Promise<T> {
-    const identities =
-      Array.isArray(identity) && identity.length > 0 ? identity : [typeof identity === "string" ? identity : ""];
-    return await this.withCaptureWriteLockHashes(
-      task,
-      identities.map((key) => (key.length === 0 ? "" : createHash("sha256").update(key).digest("hex")))
-    );
+    return await this.captureWriteLock.withCaptureWriteLock(task, identity);
   }
 
   async withCaptureWriteLockHashes<T>(task: () => Promise<T>, identityHashes: readonly string[]): Promise<T> {
-    const lockPaths = [...new Set(identityHashes.map((hash) => this.captureWriteLockPathForHash(hash)))].sort();
-    const run = async (index: number): Promise<T> => {
-      if (index === lockPaths.length) return await task();
-      return await this.withSingleCaptureWriteLock(lockPaths[index], () => run(index + 1));
-    };
-    return await run(0);
+    return await this.captureWriteLock.withCaptureWriteLockHashes(task, identityHashes);
   }
 
   private async reload(): Promise<ContentHashIndex> {
@@ -727,730 +551,5 @@ export class TombstoneBlockedCaptureIndex {
     const relative = path.relative(path.resolve(this.options.memoryDir), path.resolve(filePath));
     const category = relative.split(path.sep)[0];
     return category === "cold" || RECALL_FALLBACK_DIRS.includes(category);
-  }
-}
-
-export abstract class TombstoneBlockedCaptureIndexHost {
-  private tombstoneBlockedCaptureIndex: TombstoneBlockedCaptureIndex | null = null;
-  private readonly captureWriteLockContext = new AsyncLocalStorage<ReadonlySet<string>>();
-  protected abstract tombstoneBlockedCaptureIndexOptions(): TombstoneBlockedCaptureIndexOptions;
-  protected abstract writeStorageSecureFile(
-    filePath: string,
-    content: string | Buffer,
-    forceEncrypt?: boolean
-  ): Promise<void>;
-  protected abstract assertManagedStoragePath(filePath: string, method: string): string;
-  protected abstract notifyCatalogWriteForPath(filePath: string): void;
-  protected abstract readMemoryByPath(filePath: string): Promise<MemoryFile | null>;
-  protected abstract invalidateAllMemoriesCache(): void;
-  protected abstract invalidateKnowledgeIndexCache(): void;
-  protected abstract invalidateColdMemoriesCache(): void;
-  protected abstract bumpArtifactWriteVersion(): number;
-  protected abstract bumpMemoryStatusVersion(): void;
-  protected abstract markFactHashIndexNotAuthoritative(): void;
-
-  private async writeStorageSecureFileChunks(filePath: string, chunks: AsyncIterable<Buffer>): Promise<void> {
-    const options = this.tombstoneBlockedCaptureIndexOptions();
-    await writeMaybeEncryptedFileFromChunks(
-      filePath,
-      chunks,
-      options.secureStoreWriteKeyProvider(),
-      {},
-      options.memoryDir
-    );
-    this.notifyCatalogWriteForPath(filePath);
-  }
-
-  protected getTombstoneBlockedCaptureIndex(): TombstoneBlockedCaptureIndex {
-    if (this.tombstoneBlockedCaptureIndex === null) {
-      this.tombstoneBlockedCaptureIndex = new TombstoneBlockedCaptureIndex(this.tombstoneBlockedCaptureIndexOptions());
-    }
-    return this.tombstoneBlockedCaptureIndex;
-  }
-  async withTombstoneBlockedCaptureWriteLock<T>(
-    task: () => Promise<T>,
-    identity?: string | readonly string[]
-  ): Promise<T> {
-    const identities =
-      Array.isArray(identity) && identity.length > 0 ? identity : [typeof identity === "string" ? identity : ""];
-    const current = this.captureWriteLockContext.getStore();
-    const hashToken = (key: string): string =>
-      `\u0000${key.length === 0 ? "" : createHash("sha256").update(key).digest("hex")}`;
-    const physicalMissing = current
-      ? identities.filter((key) => !current.has(key) && !current.has(hashToken(key)))
-      : identities;
-    const next = new Set(current);
-    for (const key of identities) {
-      next.add(key);
-      next.add(hashToken(key));
-    }
-    const run = () => this.captureWriteLockContext.run(next, task);
-    if (physicalMissing.length === 0) return await run();
-    return await this.getTombstoneBlockedCaptureIndex().withCaptureWriteLock(run, physicalMissing);
-  }
-
-  protected async withTombstoneBlockedCaptureWriteLockHashes<T>(
-    task: () => Promise<T>,
-    identityHashes: readonly string[]
-  ): Promise<T> {
-    const hashes = [...new Set(identityHashes)];
-    const current = this.captureWriteLockContext.getStore();
-    const token = (hash: string): string => `\u0000${hash}`;
-    const missing = current ? hashes.filter((hash) => !current.has(token(hash))) : hashes;
-    const next = new Set(current);
-    for (const hash of hashes) next.add(token(hash));
-    const run = () => this.captureWriteLockContext.run(next, task);
-    if (missing.length === 0) return await run();
-    return await this.getTombstoneBlockedCaptureIndex().withCaptureWriteLockHashes(run, missing);
-  }
-  private async writeTombstoneBlockedMutation(
-    blocked: boolean,
-    pathname: string,
-    fileContent: string,
-    identity: string | readonly string[],
-    updateIndex: (rebuildMarker?: string, current?: MemoryFile | null) => Promise<void>,
-    beforeIndexUpdate?: () => Promise<void>,
-    coordinate = false
-  ): Promise<void> {
-    const host: TombstoneBlockedMutationHost = {
-      readCurrent: () => this.readMemoryByPath(pathname),
-      isBlocked: (memory) => this.isTombstoneBlockedMemory(memory),
-      isQueuedReview: (memory) => isQueuedReviewMemory(memory),
-      memoryIdentity: (memory) => this.offlineSyncMemoryIdentity(memory),
-      prepareWrite: () => this.getTombstoneBlockedCaptureIndex().prepareWrite(),
-      commitWrite: (marker) => this.getTombstoneBlockedCaptureIndex().commitWrite(marker),
-      discardWrite: (marker) => this.getTombstoneBlockedCaptureIndex().discardWrite(marker),
-      markUntrusted: () => this.getTombstoneBlockedCaptureIndex().markUntrusted(),
-      writeStorageSecureFile: (target, content) => this.writeStorageSecureFile(target, content),
-      withCaptureWriteLock: (task, lockIdentity) =>
-        this.withTombstoneBlockedCaptureWriteLock(task, lockIdentity),
-      logWarning: (message) => log.warn(message),
-    };
-    await runTombstoneBlockedMutation(host, {
-      blocked,
-      pathname,
-      fileContent,
-      identity,
-      updateIndex,
-      beforeIndexUpdate,
-      coordinate,
-    });
-  }
-
-  protected async writeTombstoneBlockedMemory(
-    pathname: string,
-    fileContent: string,
-    frontmatter: MemoryFrontmatter,
-    content: string,
-    beforeIndexUpdate?: () => Promise<void>
-  ): Promise<void> {
-    await this.writeTombstoneBlockedMutation(
-      tombstoneBlocked(frontmatter),
-      pathname,
-      fileContent,
-      buildExplicitCaptureDedupKey(content, frontmatter.category, frontmatter.sourceConnector),
-      (rebuildMarker) =>
-        this.getTombstoneBlockedCaptureIndex().addWrittenMemory(pathname, frontmatter, content, rebuildMarker),
-      beforeIndexUpdate
-    );
-  }
-
-  protected async writeTombstoneBlockedChunk(
-    pathname: string,
-    fileContent: string,
-    frontmatter: MemoryFrontmatter,
-    content: string,
-    findDuplicate: () => Promise<MemoryFile | null>,
-    afterWrite: () => Promise<void>
-  ): Promise<string> {
-    return await runTombstoneBlockedChunkMutation({
-      blocked: tombstoneBlocked(frontmatter),
-      id: frontmatter.id,
-      identity: buildExplicitCaptureDedupKey(content, frontmatter.category, frontmatter.sourceConnector),
-      findDuplicate,
-      persistMemory: () => this.writeTombstoneBlockedMemory(pathname, fileContent, frontmatter, content),
-      afterWrite,
-      withCaptureWriteLock: (task, identity) => this.withTombstoneBlockedCaptureWriteLock(task, identity),
-    });
-  }
-
-  protected async writeTombstoneBlockedUpdate(
-    before: MemoryFile,
-    fileContent: string,
-    frontmatter: MemoryFrontmatter,
-    content: string,
-    beforeIndexUpdate?: () => Promise<void>
-  ): Promise<void> {
-    await this.writeTombstoneBlockedMutation(
-      tombstoneBlocked(before.frontmatter) || tombstoneBlocked(frontmatter),
-      before.path,
-      fileContent,
-      [
-        ...(tombstoneBlocked(before.frontmatter)
-          ? [
-              buildExplicitCaptureDedupKey(
-                before.content,
-                before.frontmatter.category,
-                before.frontmatter.sourceConnector
-              ),
-            ]
-          : []),
-        ...(tombstoneBlocked(frontmatter)
-          ? [buildExplicitCaptureDedupKey(content, frontmatter.category, frontmatter.sourceConnector)]
-          : []),
-        ...(isQueuedReviewMemory(before)
-          ? [
-              buildExplicitCaptureDedupKey(
-                before.content,
-                before.frontmatter.category,
-                before.frontmatter.sourceConnector
-              ),
-            ]
-          : []),
-        ...(isQueuedReviewFrontmatter(frontmatter)
-          ? [buildExplicitCaptureDedupKey(content, frontmatter.category, frontmatter.sourceConnector)]
-          : []),
-      ],
-      (rebuildMarker, current) =>
-        this.getTombstoneBlockedCaptureIndex().syncUpdatedMemory(
-          current ?? before,
-          frontmatter,
-          content,
-          rebuildMarker
-        ),
-      beforeIndexUpdate,
-      isQueuedReviewMemory(before) || isQueuedReviewFrontmatter(frontmatter)
-    );
-  }
-
-  protected async writeTombstoneBlockedFrontmatter(
-    before: MemoryFile,
-    fileContent: string,
-    frontmatter: MemoryFrontmatter,
-    beforeIndexUpdate?: () => Promise<void>
-  ): Promise<void> {
-    await this.writeTombstoneBlockedMutation(
-      tombstoneBlocked(before.frontmatter) || tombstoneBlocked(frontmatter),
-      before.path,
-      fileContent,
-      [
-        ...(tombstoneBlocked(before.frontmatter)
-          ? [
-              buildExplicitCaptureDedupKey(
-                before.content,
-                before.frontmatter.category,
-                before.frontmatter.sourceConnector
-              ),
-            ]
-          : []),
-        ...(tombstoneBlocked(frontmatter)
-          ? [buildExplicitCaptureDedupKey(before.content, frontmatter.category, frontmatter.sourceConnector)]
-          : []),
-        ...(isQueuedReviewMemory(before)
-          ? [
-              buildExplicitCaptureDedupKey(
-                before.content,
-                before.frontmatter.category,
-                before.frontmatter.sourceConnector
-              ),
-            ]
-          : []),
-        ...(isQueuedReviewFrontmatter(frontmatter)
-          ? [buildExplicitCaptureDedupKey(before.content, frontmatter.category, frontmatter.sourceConnector)]
-          : []),
-      ],
-      (rebuildMarker, current) =>
-        this.getTombstoneBlockedCaptureIndex().syncUpdatedFrontmatter(
-          current ?? before,
-          frontmatter,
-          rebuildMarker
-        ),
-      beforeIndexUpdate,
-      isQueuedReviewMemory(before) || isQueuedReviewFrontmatter(frontmatter)
-    );
-  }
-
-  private isTombstoneBlockedMemory(memory: MemoryFile | null): memory is MemoryFile {
-    return memory !== null && memory.frontmatter.status === "pending_review" && Boolean(memory.frontmatter.blockedBy);
-  }
-
-  private offlineSyncMemoryIdentity(memory: MemoryFile): string {
-    return buildExplicitCaptureDedupKey(
-      memory.content,
-      memory.frontmatter.category,
-      memory.frontmatter.sourceConnector
-    );
-  }
-  private async isDuplicateTombstoneBlockedOfflineSync(
-    target: string,
-    incoming: MemoryFile,
-    streamedIdentity?: StreamedOfflineSyncIdentity
-  ): Promise<boolean> {
-    const before = await this.readMemoryByPath(target);
-    const duplicate = streamedIdentity
-      ? true
-      : await this.getTombstoneBlockedCaptureIndex().has(
-          incoming.content,
-          incoming.frontmatter.category,
-          incoming.frontmatter.sourceConnector
-        );
-    if (!duplicate) return false;
-    const incomingIdentity = streamedIdentity?.identityHash ?? this.offlineSyncMemoryIdentity(incoming);
-    const options = this.tombstoneBlockedCaptureIndexOptions();
-    const existing = [...(await options.readAllMemories()), ...(await options.readAllColdMemories())];
-    const exactMatch = existing.some(
-      (memory) =>
-        this.isTombstoneBlockedMemory(memory) &&
-        (streamedIdentity ? captureIdentityHash(memory) : this.offlineSyncMemoryIdentity(memory)) === incomingIdentity
-    );
-    if (!exactMatch) return false;
-    if (before === null) return true;
-    if (!this.isTombstoneBlockedMemory(before)) return false;
-    const beforeIdentity = streamedIdentity ? captureIdentityHash(before) : this.offlineSyncMemoryIdentity(before);
-    if (beforeIdentity !== incomingIdentity) return false;
-    return streamedIdentity
-      ? isDeepStrictEqual(before.frontmatter, incoming.frontmatter)
-      : before.content === incoming.content && isDeepStrictEqual(before.frontmatter, incoming.frontmatter);
-  }
-
-  private async invalidateAfterOfflineSyncMutation(filePath: string, ownedMarker?: string): Promise<void> {
-    this.invalidateAllMemoriesCache();
-    this.invalidateKnowledgeIndexCache();
-    this.markFactHashIndexNotAuthoritative();
-    if (filePath.includes(`${path.sep}cold${path.sep}`)) {
-      this.invalidateColdMemoriesCache();
-    }
-    await this.rebuildTombstoneBlockedCaptureAfterInvalidationForPath(filePath, ownedMarker);
-    if (filePath.includes(`${path.sep}artifacts${path.sep}`)) {
-      this.bumpArtifactWriteVersion();
-    }
-    this.bumpMemoryStatusVersion();
-  }
-
-  protected async runTombstoneBlockedArchive(
-    memory: MemoryFile,
-    task: (current: MemoryFile, markDurable: () => void) => Promise<string | null>
-  ): Promise<string | null> {
-    if (!this.isTombstoneBlockedMemory(memory) && !isQueuedReviewMemory(memory)) {
-      return await task(memory, () => {});
-    }
-    let archivedPath: string | null = null;
-    const archived = await this.runTombstoneBlockedInvalidation(
-      memory,
-      async (current, rebuildMarker, markDurable) => {
-        let archiveDurable = false;
-        const markArchiveDurable = () => {
-          archiveDurable = true;
-          markDurable();
-        };
-        archivedPath = await task(current, markArchiveDurable);
-        if (rebuildMarker !== undefined && (archivedPath !== null || archiveDurable)) {
-          await this.rebuildTombstoneBlockedCaptureAfterInvalidation(rebuildMarker);
-        } else if (rebuildMarker !== undefined) {
-          const index = this.getTombstoneBlockedCaptureIndex();
-          try { await index.discardWrite(rebuildMarker); } catch { index.markUntrusted(); }
-        }
-        return archivedPath !== null;
-      }
-    );
-    return archived ? archivedPath : null;
-  }
-
-  protected async runTombstoneBlockedInvalidation(
-    memory: MemoryFile,
-    task: (
-      current: MemoryFile,
-      rebuildMarker: string | undefined,
-      markDurable: () => void
-    ) => Promise<boolean>
-  ): Promise<boolean> {
-    let lockIdentity =
-      this.isTombstoneBlockedMemory(memory) || isQueuedReviewMemory(memory)
-        ? this.offlineSyncMemoryIdentity(memory)
-        : undefined;
-    for (;;) {
-      const attempt = async () => {
-        const options = this.tombstoneBlockedCaptureIndexOptions();
-        const current =
-          (await options.readAllMemories()).find((candidate) => candidate.frontmatter.id === memory.frontmatter.id) ??
-          (await this.readMemoryByPath(memory.path)) ??
-          (await options.readAllColdMemories()).find((candidate) => candidate.frontmatter.id === memory.frontmatter.id);
-        if (!current) return { result: false };
-        const blocked = this.isTombstoneBlockedMemory(current);
-        const queuedReview = isQueuedReviewMemory(current);
-        if (blocked || queuedReview) {
-          const currentIdentity = this.offlineSyncMemoryIdentity(current);
-          if (lockIdentity !== currentIdentity) return { retryIdentity: currentIdentity };
-        }
-        const blockedIndex = blocked ? this.getTombstoneBlockedCaptureIndex() : null;
-        let rebuildMarker: string | undefined;
-        let durable = false;
-        try {
-          if (blockedIndex) rebuildMarker = await blockedIndex.prepareWrite();
-          const result = await task(current, rebuildMarker, () => {
-            durable = true;
-          });
-          return { result };
-        } catch {
-          if (rebuildMarker && !durable && blockedIndex) {
-            try {
-              await blockedIndex.discardWrite(rebuildMarker);
-            } catch {
-              blockedIndex.markUntrusted();
-            }
-          }
-          return { result: false };
-        }
-      };
-      const result =
-        lockIdentity === undefined
-          ? await attempt()
-          : await this.withTombstoneBlockedCaptureWriteLock(attempt, lockIdentity);
-      if (result.retryIdentity !== undefined) {
-        lockIdentity = result.retryIdentity;
-        continue;
-      }
-      return result.result;
-    }
-  }
-
-  protected async runTombstoneBlockedOfflineSyncMutation(
-    target: string,
-    after: MemoryFile | null,
-    write: () => Promise<void>,
-    coordinate = false,
-    streamedIdentity?: StreamedOfflineSyncIdentity
-  ): Promise<void> {
-    for (;;) {
-      const before = await this.readMemoryByPath(target);
-      const targetIdentity = this.isTombstoneBlockedMemory(before)
-        ? this.offlineSyncMemoryIdentity(before)
-        : undefined;
-      const blocked = coordinate || targetIdentity !== undefined || this.isTombstoneBlockedMemory(after);
-      const identities = [
-        ...(targetIdentity === undefined ? [] : [targetIdentity]),
-        ...(!streamedIdentity && this.isTombstoneBlockedMemory(after)
-          ? [this.offlineSyncMemoryIdentity(after)]
-          : []),
-        ...(coordinate ? [target] : []),
-        ...(coordinate && after === null && streamedIdentity === undefined ? [""] : []),
-      ];
-      let retryIdentity: string | undefined;
-      const mutate = async (): Promise<void> => {
-        const current = await this.readMemoryByPath(target);
-        if (this.isTombstoneBlockedMemory(current)) {
-          const currentIdentity = this.offlineSyncMemoryIdentity(current);
-          if (currentIdentity !== targetIdentity) {
-            retryIdentity = currentIdentity;
-            return;
-          }
-        }
-        if (
-          this.isTombstoneBlockedMemory(after) &&
-          (await this.isDuplicateTombstoneBlockedOfflineSync(target, after, streamedIdentity))
-        ) {
-          return;
-        }
-        const marker = blocked ? await this.getTombstoneBlockedCaptureIndex().prepareWrite() : undefined;
-        let durable = false;
-        try {
-          await write();
-          durable = true;
-          if (marker) {
-            try {
-              await this.getTombstoneBlockedCaptureIndex().commitWrite(marker);
-            } catch (err) {
-              this.getTombstoneBlockedCaptureIndex().markUntrusted();
-              log.warn(`storage.offlineSyncFile committed write marker failed: ${err}`);
-            }
-          }
-          await this.invalidateAfterOfflineSyncMutation(target, marker);
-        } catch (err) {
-          if (marker && !durable) {
-            try {
-              await this.getTombstoneBlockedCaptureIndex().discardWrite(marker);
-            } catch (cleanupError) {
-              this.getTombstoneBlockedCaptureIndex().markUntrusted();
-              log.warn(`storage.offlineSyncFile failed to clear write marker: ${cleanupError}`);
-            }
-          }
-          throw err;
-        }
-      };
-      if (!blocked) {
-        await mutate();
-      } else {
-        const lockMutation = streamedIdentity
-          ? () => this.withTombstoneBlockedCaptureWriteLockHashes(mutate, [streamedIdentity.identityHash])
-          : mutate;
-        await this.withTombstoneBlockedCaptureWriteLock(lockMutation, identities);
-      }
-      if (retryIdentity !== undefined) continue;
-      return;
-    }
-  }
-
-  protected async writeTombstoneBlockedOfflineSyncFile(target: string, content: Buffer): Promise<void> {
-    await this.runTombstoneBlockedOfflineSyncMutation(
-      target,
-      this.tombstoneBlockedCaptureIndexOptions().parseMemory?.(target, content) ?? null,
-      () => this.writeStorageSecureFile(target, content)
-    );
-  }
-
-  private async prepareTombstoneBlockedOfflineSyncChunks(
-    target: string,
-    chunks: AsyncIterable<Buffer>
-  ): Promise<PreparedOfflineSyncChunks> {
-    const parseMemory = this.tombstoneBlockedCaptureIndexOptions().parseMemory;
-    if (!parseMemory) return { after: null, chunks };
-
-    const iterator = chunks[Symbol.asyncIterator]();
-    const buffered: Buffer[] = [];
-    let bufferedBytes = 0;
-    let stagePath: string | undefined;
-    let stageHandle: Awaited<ReturnType<typeof openFile>> | null = null;
-    const writeStageChunk = async (chunk: Buffer): Promise<void> => {
-      if (stageHandle === null) {
-        const stageDir = path.join(
-          this.tombstoneBlockedCaptureIndexOptions().stateDir,
-          "tombstone-blocked-capture",
-          "offline-sync-staging"
-        );
-        await mkdir(stageDir, { recursive: true });
-        stagePath = path.join(stageDir, `${randomUUID()}.stage`);
-        stageHandle = await openFile(stagePath, "w", 0o600);
-        for (const prior of buffered) await stageHandle.write(prior);
-        buffered.length = 0;
-      }
-      await stageHandle.write(chunk);
-    };
-
-    try {
-      for (;;) {
-        const next = await iterator.next();
-        if (next.done) break;
-        bufferedBytes += next.value.length;
-        if (stageHandle !== null || bufferedBytes > OFFLINE_SYNC_BUFFER_LIMIT_BYTES) {
-          await writeStageChunk(next.value);
-        } else {
-          buffered.push(next.value);
-        }
-      }
-    } catch (err) {
-      const failedStageHandle = stageHandle as Awaited<ReturnType<typeof openFile>> | null;
-      if (failedStageHandle !== null) await failedStageHandle.close().catch(() => {});
-      if (stagePath !== undefined) await unlink(stagePath).catch(() => {});
-      throw err;
-    }
-    const completedStageHandle = stageHandle as Awaited<ReturnType<typeof openFile>> | null;
-    if (completedStageHandle !== null) await completedStageHandle.close();
-    if (stagePath === undefined) {
-      let after: MemoryFile | null = null;
-      try {
-        after = parseMemory(target, Buffer.concat(buffered));
-      } catch {
-        after = null;
-      }
-      const replay = (async function* (): AsyncIterable<Buffer> {
-        for (const chunk of buffered) yield chunk;
-      })();
-      return { after, chunks: replay };
-    }
-    try {
-      return await this.describeStagedOfflineSyncChunks(target, stagePath, parseMemory);
-    } catch (err) {
-      await unlink(stagePath).catch(() => {});
-      throw err;
-    }
-  }
-
-  private async describeStagedOfflineSyncChunks(
-    target: string,
-    stagePath: string,
-    parseMemory: OfflineSyncMemoryParser
-  ): Promise<PreparedOfflineSyncChunks> {
-    const handle = await openFile(stagePath, "r");
-    const prefix = Buffer.alloc(OFFLINE_SYNC_FRONTMATTER_LIMIT_BYTES);
-    let bytesRead = 0;
-    try {
-      ({ bytesRead } = await handle.read(prefix, 0, prefix.length, 0));
-    } finally {
-      await handle.close();
-    }
-    const delimiter = Buffer.from("\n---\n", "utf8");
-    const delimiterStart = prefix.subarray(0, bytesRead).indexOf(delimiter, 4);
-    if (delimiterStart < 0) {
-      return { after: null, chunks: this.readStagedOfflineSyncChunks(stagePath), stagePath };
-    }
-    const bodyOffset = delimiterStart + delimiter.length;
-    let after: MemoryFile | null = null;
-    try {
-      after = parseMemory(target, prefix.subarray(0, bodyOffset));
-    } catch {
-      after = null;
-    }
-    if (after === null) {
-      return { after: null, chunks: this.readStagedOfflineSyncChunks(stagePath), stagePath };
-    }
-    const streamedIdentity = await this.computeStagedOfflineSyncIdentity(stagePath, bodyOffset, after);
-    return {
-      after,
-      chunks: this.readStagedOfflineSyncChunks(stagePath),
-      stagePath,
-      streamedIdentity,
-    };
-  }
-
-  private async computeStagedOfflineSyncIdentity(
-    stagePath: string,
-    bodyOffset: number,
-    after: MemoryFile
-  ): Promise<StreamedOfflineSyncIdentity> {
-    let normalizedLength = 0;
-    const countNormalizer = createExplicitContentNormalizer((chunk) => {
-      normalizedLength += chunk.length;
-    });
-    await forEachStagedTextChunk(stagePath, bodyOffset, (text) => countNormalizer.push(text));
-    countNormalizer.finish();
-
-    const category = after.frontmatter.category;
-    const connector = after.frontmatter.sourceConnector?.trim() ?? "";
-    const encode = (label: string, value: string): string => `${label} ${value.length} ${value}`;
-    const identityPrefix =
-      [encode("category", category), encode("connector", connector), `content ${normalizedLength} `].join(" ");
-    const identityHash = createHash("sha256");
-    identityHash.update(identityPrefix);
-    const identityNormalizer = createExplicitContentNormalizer((chunk) => identityHash.update(chunk));
-    await forEachStagedTextChunk(stagePath, bodyOffset, (text) => identityNormalizer.push(text));
-    identityNormalizer.finish();
-    return {
-      identityHash: identityHash.digest("hex"),
-    };
-  }
-
-  private async *readStagedOfflineSyncChunks(stagePath: string): AsyncIterable<Buffer> {
-    const handle = await openFile(stagePath, "r");
-    const buffer = Buffer.alloc(OFFLINE_SYNC_STAGE_READ_BYTES);
-    try {
-      for (;;) {
-        const { bytesRead } = await handle.read(buffer, 0, buffer.length, null);
-        if (bytesRead === 0) break;
-        yield Buffer.from(buffer.subarray(0, bytesRead));
-      }
-    } finally {
-      await handle.close();
-    }
-  }
-
-  protected async writeTombstoneBlockedOfflineSyncFileChunks(
-    target: string,
-    chunks: AsyncIterable<Buffer>
-  ): Promise<void> {
-    const coordinate = this.getTombstoneBlockedCaptureIndex().shouldRebuildAfterInvalidationForPath(target);
-    const prepared: PreparedOfflineSyncChunks = coordinate
-      ? await this.prepareTombstoneBlockedOfflineSyncChunks(target, chunks)
-      : { after: null, chunks };
-    const incoming = this.isTombstoneBlockedMemory(prepared.after) ? prepared.after : null;
-    const streamedIdentity = prepared.streamedIdentity;
-    const incomingIdentity =
-      streamedIdentity?.identityHash ?? (incoming ? this.offlineSyncMemoryIdentity(incoming) : null);
-    const mutate = async (): Promise<void> => {
-      try {
-        await this.runTombstoneBlockedOfflineSyncMutation(
-          target,
-          prepared.after,
-          () => this.writeStorageSecureFileChunks(target, prepared.chunks),
-          coordinate,
-          streamedIdentity
-        );
-      } finally {
-        if (prepared.stagePath !== undefined) await unlink(prepared.stagePath).catch(() => {});
-      }
-    };
-    if (!coordinate) {
-      await mutate();
-      return;
-    }
-    for (;;) {
-      const before = await this.readMemoryByPath(target);
-      const targetIdentity = this.isTombstoneBlockedMemory(before)
-        ? this.offlineSyncMemoryIdentity(before)
-        : undefined;
-      const identities = [
-        ...(!streamedIdentity && incomingIdentity ? [incomingIdentity] : []),
-        ...(targetIdentity ? [targetIdentity] : []),
-        target,
-        ...(prepared.after === null && streamedIdentity === undefined ? [""] : []),
-      ];
-      let retryIdentity: string | undefined;
-      const checkAndMutate = async (): Promise<void> => {
-        const current = await this.readMemoryByPath(target);
-        if (this.isTombstoneBlockedMemory(current)) {
-          const currentIdentity = this.offlineSyncMemoryIdentity(current);
-          if (currentIdentity !== targetIdentity) {
-            retryIdentity = currentIdentity;
-            return;
-          }
-        }
-        await mutate();
-      };
-      const lockTask = streamedIdentity
-        ? () => this.withTombstoneBlockedCaptureWriteLockHashes(checkAndMutate, [streamedIdentity.identityHash])
-        : checkAndMutate;
-      await this.withTombstoneBlockedCaptureWriteLock(lockTask, identities);
-      if (retryIdentity === undefined) return;
-    }
-  }
-
-  async writeOfflineSyncFile(filePath: string, content: Buffer): Promise<void> {
-    const target = this.assertManagedStoragePath(filePath, "storage.writeOfflineSyncFile");
-    await this.writeTombstoneBlockedOfflineSyncFile(target, content);
-  }
-
-  async writeOfflineSyncFileChunks(filePath: string, chunks: AsyncIterable<Buffer>): Promise<void> {
-    const target = this.assertManagedStoragePath(filePath, "storage.writeOfflineSyncFileChunks");
-    await this.writeTombstoneBlockedOfflineSyncFileChunks(target, chunks);
-  }
-
-  async deleteOfflineSyncFile(filePath: string): Promise<void> {
-    const target = this.assertManagedStoragePath(filePath, "storage.deleteOfflineSyncFile");
-    await this.deleteTombstoneBlockedOfflineSyncFile(target);
-  }
-
-  protected async deleteTombstoneBlockedOfflineSyncFile(target: string): Promise<void> {
-    await this.runTombstoneBlockedOfflineSyncMutation(target, null, async () => {
-      await unlink(target).catch((error: unknown) => {
-        if (isErrnoCode(error, "ENOENT")) return;
-        throw error;
-      });
-    });
-  }
-
-  async checkTombstoneBlockedExplicitCapture(
-    ...args: [string, string, string?]
-  ): Promise<{ has: boolean; authoritative: boolean }> {
-    return this.getTombstoneBlockedCaptureIndex().check(...args);
-  }
-
-  async hasTombstoneBlockedExplicitCapture(...args: [string, string, string?]): Promise<boolean> {
-    return this.getTombstoneBlockedCaptureIndex().has(...args);
-  }
-
-  async isTombstoneBlockedExplicitCaptureIndexAuthoritative(): Promise<boolean> {
-    return this.getTombstoneBlockedCaptureIndex().isAuthoritative();
-  }
-
-  protected async rebuildTombstoneBlockedCaptureAfterInvalidationForPath(
-    filePath: string,
-    ownedMarker?: string
-  ): Promise<void> {
-    const index = this.getTombstoneBlockedCaptureIndex();
-    if (index.shouldRebuildAfterInvalidationForPath(filePath)) {
-      await this.rebuildTombstoneBlockedCaptureAfterInvalidation(ownedMarker);
-    }
-  }
-  protected async rebuildTombstoneBlockedCaptureAfterInvalidation(ownedMarker?: string): Promise<void> {
-    await this.getTombstoneBlockedCaptureIndex().rebuildAfterInvalidation(ownedMarker);
   }
 }
